+++ lvgl.h
@@ -77,7 +77,7 @@ extern "C" {
  *********************/
 /*Current version of LVGL*/
 #define LVGL_VERSION_MAJOR   7
-#define LVGL_VERSION_MINOR   2
+#define LVGL_VERSION_MINOR   3
 #define LVGL_VERSION_PATCH   0
 #define LVGL_VERSION_INFO ""