+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## v7.3.0 (planned on 04.08.2020)
 Available in the `dev` branch
 
+### New features
+- Add `lv_task_get_next`
+- Add `lv_event_send_refresh`, `lv_event_send_refresh_recursive` to easily send `LV_EVENT_REFRESH` to object
+- Add `lv_tabview_set_tab_name()` function - used to change a tab's name
+
 ## v7.2.0 (planned on 21.07.2020)
 *Available in the `master` branch*